# Conflicts:
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/event/actor/ActorEvent.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/event/actorinstancerelationship/ActorInstanceRelationshipEvent.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/event/instance/InstanceEvent.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/event/post/PostEvent.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/actor/Actor2.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/actor/ActorId.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/actor/ActorName.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/actor/ActorPostsCount.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/actorinstancerelationship/ActorInstanceRelationship.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/instance/Instance.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/post/Post2.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/service/actor/RemoteActorCheckDomainService.kt
#	hideout-core/src/main/kotlin/dev/usbharu/hideout/core/infrastructure/factory/Actor2FactoryImpl.kt

+++ hideout-core/src/main/kotlin/dev/usbharu/hideout/core/domain/model/actor/ActorPostsCount.kt
@@ -17,7 +17,7 @@
 package dev.usbharu.hideout.core.domain.model.actor
 
 @JvmInline
-value class ActorPostsCount(private val postsCount: Long) {
+value class ActorPostsCount(private val postsCount: Int) {
     init {
         require(0 <= this.postsCount) { "Posts count must be greater than 0" }
     }
